requirements.txt
flask
pydantic
fastapi
transformers
torch
python-dotenv
hf_xet
sentencepiece
protobuf